refactor: graph drawers moved to Cairo or Matplotlib submodule
