+++ src/igraph/drawing/graph.py
@@ -13,526 +13,11 @@ L{CytoscapeGraphDrawer}. L{CytoscapeGraphDrawer} can also fetch the current
 network from Cytoscape and convert it to igraph format.
 """
 
-from abc import abstractmethod
-from math import atan2, cos, pi, sin, tan
 from warnings import warn
 
-from igraph._igraph import convex_hull, VertexSeq
-from igraph.configuration import Configuration
-from igraph.drawing.baseclasses import AbstractDrawer, AbstractXMLRPCDrawer
-from igraph.drawing.cairo.base import AbstractCairoDrawer
-from igraph.drawing.cairo.edge import ArrowEdgeDrawer
-from igraph.drawing.cairo.text import CairoTextDrawer
-from igraph.drawing.cairo.vertex import CairoVertexDrawer
-from igraph.drawing.cairo.utils import find_cairo
-from igraph.drawing.matplotlib.edge import MatplotlibEdgeDrawer
-from igraph.drawing.matplotlib.utils import find_matplotlib
-from igraph.drawing.matplotlib.vertex import MatplotlibVertexDrawer
-from igraph.drawing.text import TextAlignment
-from igraph.drawing.shapes import CairoPolygonDrawer, MatplotlibPolygonDrawer
-from igraph.drawing.utils import Point
-from igraph.layout import Layout
+from igraph.drawing.baseclasses import AbstractGraphDrawer, AbstractXMLRPCDrawer
 
-__all__ = (
-    "CairoGraphDrawer",
-    "MatplotlibGraphDrawer",
-    "CytoscapeGraphDrawer",
-)
-
-cairo = find_cairo()
-mpl, plt = find_matplotlib()
-
-#####################################################################
-
-
-class AbstractGraphDrawer(AbstractDrawer):
-    """Abstract class that serves as a base class for anything that
-    draws an igraph.Graph."""
-
-    @abstractmethod
-    def draw(self, graph, *args, **kwds):
-        """Abstract method, must be implemented in derived classes."""
-        raise NotImplementedError
-
-    def ensure_layout(self, layout, graph=None):
-        """Helper method that ensures that I{layout} is an instance
-        of L{Layout}. If it is not, the method will try to convert
-        it to a L{Layout} according to the following rules:
-
-          - If I{layout} is a string, it is assumed to be a name
-            of an igraph layout, and it will be passed on to the
-            C{layout} method of the given I{graph} if I{graph} is
-            not C{None}.
-
-          - If I{layout} is C{None}, the C{layout} method of
-            I{graph} will be invoked with no parameters, which
-            will call the default layout algorithm.
-
-          - Otherwise, I{layout} will be passed on to the constructor
-            of L{Layout}. This handles lists of lists, lists of tuples
-            and such.
-
-        If I{layout} is already a L{Layout} instance, it will still
-        be copied and a copy will be returned. This is because graph
-        drawers are allowed to transform the layout for their purposes,
-        and we don't want the transformation to propagate back to the
-        caller.
-        """
-        if isinstance(layout, Layout):
-            layout = Layout(layout.coords)
-        elif isinstance(layout, str) or layout is None:
-            layout = graph.layout(layout)
-        else:
-            layout = Layout(layout)
-        return layout
-
-    @staticmethod
-    def _determine_edge_order(graph, kwds):
-        """Returns the order in which the edge of the given graph have to be
-        drawn, assuming that the relevant keyword arguments (C{edge_order} and
-        C{edge_order_by}) are given in C{kwds} as a dictionary. If neither
-        C{edge_order} nor C{edge_order_by} is present in C{kwds}, this
-        function returns C{None} to indicate that the graph drawer is free to
-        choose the most convenient edge ordering."""
-        if "edge_order" in kwds:
-            # Edge order specified explicitly
-            return kwds["edge_order"]
-
-        if kwds.get("edge_order_by") is None:
-            # No edge order specified
-            return None
-
-        # Order edges by the value of some attribute
-        edge_order_by = kwds["edge_order_by"]
-        reverse = False
-        if isinstance(edge_order_by, tuple):
-            edge_order_by, reverse = edge_order_by
-            if isinstance(reverse, str):
-                reverse = reverse.lower().startswith("desc")
-        attrs = graph.es[edge_order_by]
-        edge_order = sorted(
-            list(range(len(attrs))), key=attrs.__getitem__, reverse=bool(reverse)
-        )
-
-        return edge_order
-
-    @staticmethod
-    def _determine_vertex_order(graph, kwds):
-        """Returns the order in which the vertices of the given graph have to be
-        drawn, assuming that the relevant keyword arguments (C{vertex_order} and
-        C{vertex_order_by}) are given in C{kwds} as a dictionary. If neither
-        C{vertex_order} nor C{vertex_order_by} is present in C{kwds}, this
-        function returns C{None} to indicate that the graph drawer is free to
-        choose the most convenient vertex ordering."""
-        if "vertex_order" in kwds:
-            # Vertex order specified explicitly
-            return kwds["vertex_order"]
-
-        if kwds.get("vertex_order_by") is None:
-            # No vertex order specified
-            return None
-
-        # Order vertices by the value of some attribute
-        vertex_order_by = kwds["vertex_order_by"]
-        reverse = False
-        if isinstance(vertex_order_by, tuple):
-            vertex_order_by, reverse = vertex_order_by
-            if isinstance(reverse, str):
-                reverse = reverse.lower().startswith("desc")
-        attrs = graph.vs[vertex_order_by]
-        vertex_order = sorted(
-            list(range(len(attrs))), key=attrs.__getitem__, reverse=bool(reverse)
-        )
-
-        return vertex_order
-
-
-#####################################################################
-
-
-class AbstractCairoGraphDrawer(AbstractGraphDrawer, AbstractCairoDrawer):
-    """Abstract base class for graph drawers that draw on a Cairo canvas."""
-
-    def __init__(self, context, bbox):
-        """Constructs the graph drawer and associates it to the given
-        Cairo context and the given L{BoundingBox}.
-
-        @param context: the context on which we will draw
-        @param bbox:    the bounding box within which we will draw.
-                        Can be anything accepted by the constructor
-                        of L{BoundingBox} (i.e., a 2-tuple, a 4-tuple
-                        or a L{BoundingBox} object).
-        """
-        AbstractCairoDrawer.__init__(self, context, bbox)
-        AbstractGraphDrawer.__init__(self)
-
-
-#####################################################################
-
-
-class CairoGraphDrawer(AbstractCairoGraphDrawer):
-    """Class implementing the default visualisation of a graph.
-
-    The default visualisation of a graph draws the nodes on a 2D plane
-    according to a given L{Layout}, then draws a straight or curved
-    edge between nodes connected by edges. This is the visualisation
-    used when one invokes the L{plot()} function on a L{Graph} object.
-
-    See L{Graph.__plot__()} for the keyword arguments understood by
-    this drawer."""
-
-    def __init__(
-        self,
-        context,
-        bbox,
-        vertex_drawer_factory=CairoVertexDrawer,
-        edge_drawer_factory=ArrowEdgeDrawer,
-        label_drawer_factory=CairoTextDrawer,
-    ):
-        """Constructs the graph drawer and associates it to the given
-        Cairo context and the given L{BoundingBox}.
-
-        @param context: the context on which we will draw
-        @param bbox:    the bounding box within which we will draw.
-                        Can be anything accepted by the constructor
-                        of L{BoundingBox} (i.e., a 2-tuple, a 4-tuple
-                        or a L{BoundingBox} object).
-        @param vertex_drawer_factory: a factory method that returns an
-                        L{AbstractCairoVertexDrawer} instance bound to a
-                        given Cairo context. The factory method must take
-                        four parameters: the Cairo context, the bounding
-                        box of the drawing area, the palette to be
-                        used for drawing colored vertices, and the graph layout.
-                        The default vertex drawer is L{DefaultVertexDrawer}.
-        @param edge_drawer_factory: a factory method that returns an
-                        L{AbstractEdgeDrawer} instance bound to a
-                        given Cairo context. The factory method must take
-                        two parameters: the Cairo context and the palette
-                        to be used for drawing colored edges. You can use
-                        any of the actual L{AbstractEdgeDrawer}
-                        implementations here to control the style of
-                        edges drawn by igraph. The default edge drawer is
-                        L{ArrowEdgeDrawer}.
-        @param label_drawer_factory: a factory method that returns a
-                        L{CairoTextDrawer} instance bound to a given Cairo
-                        context. The method must take one parameter: the
-                        Cairo context. The default label drawer is
-                        L{CairoTextDrawer}.
-        """
-        AbstractCairoGraphDrawer.__init__(self, context, bbox)
-        self.vertex_drawer_factory = vertex_drawer_factory
-        self.edge_drawer_factory = edge_drawer_factory
-        self.label_drawer_factory = label_drawer_factory
-
-    def draw(self, graph, *args, **kwds):
-        # Positional arguments are not used
-        if args:
-            warn(
-                "Positional arguments to plot functions are ignored "
-                "and will be deprecated soon.",
-                DeprecationWarning,
-            )
-
-        # Some abbreviations for sake of simplicity
-        directed = graph.is_directed()
-        context = self.context
-
-        # Palette
-        palette = kwds.pop("palette", None)
-
-        # Calculate/get the layout of the graph
-        layout = self.ensure_layout(kwds.get("layout", None), graph)
-
-        # Determine the size of the margin on each side
-        margin = kwds.get("margin", 0)
-        try:
-            margin = list(margin)
-        except TypeError:
-            margin = [margin]
-        while len(margin) < 4:
-            margin.extend(margin)
-
-        # Contract the drawing area by the margin and fit the layout
-        bbox = self.bbox.contract(margin)
-        layout.fit_into(bbox, keep_aspect_ratio=kwds.get("keep_aspect_ratio", False))
-
-        # Decide whether we need to calculate the curvature of edges
-        # automatically -- and calculate them if needed.
-        autocurve = kwds.get("autocurve", None)
-        if autocurve or (
-            autocurve is None
-            and "edge_curved" not in kwds
-            and "curved" not in graph.edge_attributes()
-            and graph.ecount() < 10000
-        ):
-            from igraph import autocurve
-
-            default = kwds.get("edge_curved", 0)
-            if default is True:
-                default = 0.5
-            default = float(default)
-            kwds["edge_curved"] = autocurve(graph, attribute=None, default=default)
-
-        # Construct the vertex, edge and label drawers
-        vertex_drawer = self.vertex_drawer_factory(context, bbox, palette, layout)
-        edge_drawer = self.edge_drawer_factory(context, palette)
-        label_drawer = self.label_drawer_factory(context)
-
-        # Construct the visual vertex/edge builders based on the specifications
-        # provided by the vertex_drawer and the edge_drawer
-        vertex_builder = vertex_drawer.VisualVertexBuilder(graph.vs, kwds)
-        edge_builder = edge_drawer.VisualEdgeBuilder(graph.es, kwds)
-
-        # Determine the order in which we will draw the vertices and edges
-        vertex_order = self._determine_vertex_order(graph, kwds)
-        edge_order = self._determine_edge_order(graph, kwds)
-
-        # Draw the highlighted groups (if any)
-        if "mark_groups" in kwds:
-            mark_groups = kwds["mark_groups"]
-
-            # Deferred import to avoid a cycle in the import graph
-            from igraph.clustering import VertexClustering, VertexCover
-
-            # Figure out what to do with mark_groups in order to be able to
-            # iterate over it and get memberlist-color pairs
-            if isinstance(mark_groups, dict):
-                # Dictionary mapping vertex indices or tuples of vertex
-                # indices to colors
-                group_iter = iter(mark_groups.items())
-            elif isinstance(mark_groups, (VertexClustering, VertexCover)):
-                # Vertex clustering
-                group_iter = ((group, color) for color, group in enumerate(mark_groups))
-            elif hasattr(mark_groups, "__iter__"):
-                # Lists, tuples, iterators etc
-                group_iter = iter(mark_groups)
-            else:
-                # False
-                group_iter = iter({}.items())
-
-            # We will need a polygon drawer to draw the convex hulls
-            polygon_drawer = CairoPolygonDrawer(context, bbox)
-
-            # Iterate over color-memberlist pairs
-            for group, color_id in group_iter:
-                if not group or color_id is None:
-                    continue
-
-                color = palette.get(color_id)
-
-                if isinstance(group, VertexSeq):
-                    group = [vertex.index for vertex in group]
-                if not hasattr(group, "__iter__"):
-                    raise TypeError("group membership list must be iterable")
-
-                # Get the vertex indices that constitute the convex hull
-                hull = [group[i] for i in convex_hull([layout[idx] for idx in group])]
-
-                # Calculate the preferred rounding radius for the corners
-                corner_radius = 1.25 * max(vertex_builder[idx].size for idx in hull)
-
-                # Construct the polygon
-                polygon = [layout[idx] for idx in hull]
-
-                if len(polygon) == 2:
-                    # Expand the polygon (which is a flat line otherwise)
-                    a, b = Point(*polygon[0]), Point(*polygon[1])
-                    c = corner_radius * (a - b).normalized()
-                    n = Point(-c[1], c[0])
-                    polygon = [a + n, b + n, b - c, b - n, a - n, a + c]
-                else:
-                    # Expand the polygon around its center of mass
-                    center = Point(
-                        *[sum(coords) / float(len(coords)) for coords in zip(*polygon)]
-                    )
-                    polygon = [
-                        Point(*point).towards(center, -corner_radius)
-                        for point in polygon
-                    ]
-
-                # Draw the hull
-                context.set_source_rgba(color[0], color[1], color[2], color[3] * 0.25)
-                polygon_drawer.draw_path(polygon, corner_radius=corner_radius)
-                context.fill_preserve()
-                context.set_source_rgba(*color)
-                context.stroke()
-
-        # Construct the iterator that we will use to draw the edges
-        es = graph.es
-        if edge_order is None:
-            # Default edge order
-            edge_coord_iter = zip(es, edge_builder)
-        else:
-            # Specified edge order
-            edge_coord_iter = ((es[i], edge_builder[i]) for i in edge_order)
-
-        # Draw the edges
-        if directed:
-            drawer_method = edge_drawer.draw_directed_edge
-        else:
-            drawer_method = edge_drawer.draw_undirected_edge
-        for edge, visual_edge in edge_coord_iter:
-            src, dest = edge.tuple
-            src_vertex, dest_vertex = vertex_builder[src], vertex_builder[dest]
-            drawer_method(visual_edge, src_vertex, dest_vertex)
-
-        # Construct the iterator that we will use to draw the vertices
-        vs = graph.vs
-        if vertex_order is None:
-            # Default vertex order
-            vertex_coord_iter = zip(vs, vertex_builder, layout)
-        else:
-            # Specified vertex order
-            vertex_coord_iter = (
-                (vs[i], vertex_builder[i], layout[i]) for i in vertex_order
-            )
-
-        # Draw the vertices
-        drawer_method = vertex_drawer.draw
-        context.set_line_width(1)
-        for vertex, visual_vertex, coords in vertex_coord_iter:
-            drawer_method(visual_vertex, vertex, coords)
-
-        # Decide whether the labels have to be wrapped
-        wrap = kwds.get("wrap_labels")
-        if wrap is None:
-            wrap = Configuration.instance()["plotting.wrap_labels"]
-        wrap = bool(wrap)
-
-        # Construct the iterator that we will use to draw the vertex labels
-        if vertex_order is None:
-            # Default vertex order
-            vertex_coord_iter = zip(vertex_builder, layout)
-        else:
-            # Specified vertex order
-            vertex_coord_iter = ((vertex_builder[i], layout[i]) for i in vertex_order)
-
-        # Draw the vertex labels
-        for vertex, coords in vertex_coord_iter:
-            if vertex.label is None:
-                continue
-
-            # Set the font family, size, color and text
-            context.select_font_face(
-                vertex.font, cairo.FONT_SLANT_NORMAL, cairo.FONT_WEIGHT_NORMAL
-            )
-            context.set_font_size(vertex.label_size)
-            context.set_source_rgba(*vertex.label_color)
-            label_drawer.text = vertex.label
-
-            if vertex.label_dist:
-                # Label is displaced from the center of the vertex.
-                _, yb, w, h, _, _ = label_drawer.text_extents()
-                w, h = w / 2.0, h / 2.0
-                radius = vertex.label_dist * vertex.size / 2.0
-                # First we find the reference point that is at distance `radius'
-                # from the vertex in the direction given by `label_angle'.
-                # Then we place the label in a way that the line connecting the
-                # center of the bounding box of the label with the center of the
-                # vertex goes through the reference point and the reference
-                # point lies exactly on the bounding box of the vertex.
-                alpha = vertex.label_angle % (2 * pi)
-                cx = coords[0] + radius * cos(alpha)
-                cy = coords[1] - radius * sin(alpha)
-                # Now we have the reference point. We have to decide which side
-                # of the label box will intersect with the line that connects
-                # the center of the label with the center of the vertex.
-                if w > 0:
-                    beta = atan2(h, w) % (2 * pi)
-                else:
-                    beta = pi / 2.0
-                gamma = pi - beta
-                if alpha > 2 * pi - beta or alpha <= beta:
-                    # Intersection at left edge of label
-                    cx += w
-                    cy -= tan(alpha) * w
-                elif alpha > beta and alpha <= gamma:
-                    # Intersection at bottom edge of label
-                    try:
-                        cx += h / tan(alpha)
-                    except Exception:
-                        pass  # tan(alpha) == inf
-                    cy -= h
-                elif alpha > gamma and alpha <= gamma + 2 * beta:
-                    # Intersection at right edge of label
-                    cx -= w
-                    cy += tan(alpha) * w
-                else:
-                    # Intersection at top edge of label
-                    try:
-                        cx -= h / tan(alpha)
-                    except Exception:
-                        pass  # tan(alpha) == inf
-                    cy += h
-                # Draw the label
-                label_drawer.draw_at(cx - w, cy - h - yb, wrap=wrap)
-            else:
-                # Label is exactly in the center of the vertex
-                cx, cy = coords
-                half_size = vertex.size / 2.0
-                label_drawer.bbox = (
-                    cx - half_size,
-                    cy - half_size,
-                    cx + half_size,
-                    cy + half_size,
-                )
-                label_drawer.draw(wrap=wrap)
-
-        # Construct the iterator that we will use to draw the edge labels
-        es = graph.es
-        if edge_order is None:
-            # Default edge order
-            edge_coord_iter = zip(es, edge_builder)
-        else:
-            # Specified edge order
-            edge_coord_iter = ((es[i], edge_builder[i]) for i in edge_order)
-
-        # Draw the edge labels
-        for edge, visual_edge in edge_coord_iter:
-            if visual_edge.label is None:
-                continue
-
-            # Set the font family, size, color and text
-            context.select_font_face(
-                visual_edge.font, cairo.FONT_SLANT_NORMAL, cairo.FONT_WEIGHT_NORMAL
-            )
-            context.set_font_size(visual_edge.label_size)
-            context.set_source_rgba(*visual_edge.label_color)
-            label_drawer.text = visual_edge.label
-
-            # Ask the edge drawer to propose an anchor point for the label
-            src, dest = edge.tuple
-            src_vertex, dest_vertex = vertex_builder[src], vertex_builder[dest]
-            (x, y), (halign, valign) = edge_drawer.get_label_position(
-                edge, src_vertex, dest_vertex
-            )
-
-            # Measure the text
-            _, yb, w, h, _, _ = label_drawer.text_extents()
-            w /= 2.0
-            h /= 2.0
-
-            # Place the text relative to the edge
-            if halign == TextAlignment.RIGHT:
-                x -= w
-            elif halign == TextAlignment.LEFT:
-                x += w
-            if valign == TextAlignment.BOTTOM:
-                y -= h - yb / 2.0
-            elif valign == TextAlignment.TOP:
-                y += h
-
-            # Draw the edge label
-            label_drawer.halign = halign
-            label_drawer.valign = valign
-            label_drawer.bbox = (x - w, y - h, x + w, y + h)
-            label_drawer.draw(wrap=wrap)
-
-
-# TODO: deprecate
-DefaultGraphDrawer = CairoGraphDrawer
-
-#####################################################################
+__all__ = ("CytoscapeGraphDrawer",)
 
 
 class CytoscapeGraphDrawer(AbstractXMLRPCDrawer, AbstractGraphDrawer):
@@ -871,295 +356,3 @@ class GephiGraphStreamingDrawer(AbstractGraphDrawer):
             )
 
         self.streamer.post(graph, self.connection, encoder=kwds.get("encoder"))
-
-
-#####################################################################
-
-
-class MatplotlibGraphDrawer(AbstractGraphDrawer):
-    """Graph drawer that uses a pyplot.Axes as context"""
-
-    _shape_dict = {
-        "rectangle": "s",
-        "circle": "o",
-        "hidden": "none",
-        "triangle-up": "^",
-        "triangle-down": "v",
-    }
-
-    def __init__(
-        self,
-        ax,
-        vertex_drawer_factory=MatplotlibVertexDrawer,
-        edge_drawer_factory=MatplotlibEdgeDrawer,
-    ):
-        """Constructs the graph drawer and associates it with the mpl Axes
-
-        @param ax: the matplotlib Axes to draw into.
-        @param vertex_drawer_factory: a factory method that returns an
-            L{AbstractVertexDrawer} instance bound to the given Matplotlib axes.
-            The factory method must take three parameters: the axes and the
-            palette to be used for drawing colored vertices, and the layout of
-            the graph. The default vertex drawer is L{MatplotlibVertexDrawer}.
-        @param edge_drawer_factory: a factory method that returns an
-            L{AbstractEdgeDrawer} instance bound to a given Matplotlib Axes.
-            The factory method must take two parameters: the Axes and the palette
-            to be used for drawing colored edges. The default edge drawer is
-            L{MatplotlibEdgeDrawer}.
-        """
-        self.ax = ax
-        self.vertex_drawer_factory = vertex_drawer_factory
-        self.edge_drawer_factory = edge_drawer_factory
-
-    def draw(self, graph, *args, **kwds):
-        # Deferred import to avoid a cycle in the import graph
-        from igraph.clustering import VertexClustering, VertexCover
-
-        # Positional arguments are not used
-        if args:
-            warn(
-                "Positional arguments to plot functions are ignored "
-                "and will be deprecated soon.",
-                DeprecationWarning,
-            )
-
-        # Some abbreviations for sake of simplicity
-        directed = graph.is_directed()
-        ax = self.ax
-
-        # Palette
-        palette = kwds.pop("palette", None)
-
-        # Calculate/get the layout of the graph
-        layout = self.ensure_layout(kwds.get("layout", None), graph)
-
-        # Decide whether we need to calculate the curvature of edges
-        # automatically -- and calculate them if needed.
-        autocurve = kwds.get("autocurve", None)
-        if autocurve or (
-            autocurve is None
-            and "edge_curved" not in kwds
-            and "curved" not in graph.edge_attributes()
-            and graph.ecount() < 10000
-        ):
-            from igraph import autocurve
-
-            default = kwds.get("edge_curved", 0)
-            if default is True:
-                default = 0.5
-            default = float(default)
-            kwds["edge_curved"] = autocurve(
-                graph,
-                attribute=None,
-                default=default,
-            )
-
-        # Construct the vertex, edge and label drawers
-        vertex_drawer = self.vertex_drawer_factory(
-            ax,
-            palette,
-            layout,
-        )
-        edge_drawer = self.edge_drawer_factory(ax, palette)
-
-        # Construct the visual vertex/edge builders based on the specifications
-        # provided by the vertex_drawer and the edge_drawer
-        vertex_builder = vertex_drawer.VisualVertexBuilder(graph.vs, kwds)
-        edge_builder = edge_drawer.VisualEdgeBuilder(graph.es, kwds)
-
-        # Draw the highlighted groups (if any)
-        if "mark_groups" in kwds:
-            mark_groups = kwds["mark_groups"]
-
-            # Deferred import to avoid a cycle in the import graph
-            from igraph.clustering import VertexClustering, VertexCover
-
-            # Figure out what to do with mark_groups in order to be able to
-            # iterate over it and get memberlist-color pairs
-            if isinstance(mark_groups, dict):
-                # Dictionary mapping vertex indices or tuples of vertex
-                # indices to colors
-                group_iter = iter(mark_groups.items())
-            elif isinstance(mark_groups, (VertexClustering, VertexCover)):
-                # Vertex clustering
-                group_iter = ((group, color) for color, group in enumerate(mark_groups))
-            elif hasattr(mark_groups, "__iter__"):
-                # Lists, tuples, iterators etc
-                group_iter = iter(mark_groups)
-            else:
-                # False
-                group_iter = iter({}.items())
-
-            if kwds.get("legend", False):
-                legend_info = {
-                    "handles": [],
-                    "labels": [],
-                }
-
-            # Iterate over color-memberlist pairs
-            for group, color_id in group_iter:
-                if not group or color_id is None:
-                    continue
-
-                color = palette.get(color_id)
-
-                if isinstance(group, VertexSeq):
-                    group = [vertex.index for vertex in group]
-                if not hasattr(group, "__iter__"):
-                    raise TypeError("group membership list must be iterable")
-
-                # Get the vertex indices that constitute the convex hull
-                hull = [group[i] for i in convex_hull([layout[idx] for idx in group])]
-
-                # Calculate the preferred rounding radius for the corners
-                corner_radius = 1.25 * max(vertex_builder[idx].size for idx in hull)
-
-                # Construct the polygon
-                polygon = [layout[idx] for idx in hull]
-
-                if len(polygon) == 2:
-                    # Expand the polygon (which is a flat line otherwise)
-                    a, b = Point(*polygon[0]), Point(*polygon[1])
-                    c = corner_radius * (a - b).normalized()
-                    n = Point(-c[1], c[0])
-                    polygon = [a + n, b + n, b - c, b - n, a - n, a + c]
-                else:
-                    # Expand the polygon around its center of mass
-                    center = Point(
-                        *[sum(coords) / float(len(coords)) for coords in zip(*polygon)]
-                    )
-                    polygon = [
-                        Point(*point).towards(center, -corner_radius)
-                        for point in polygon
-                    ]
-
-                # Draw the hull
-                facecolor = (color[0], color[1], color[2], 0.25 * color[3])
-                drawer = MatplotlibPolygonDrawer(ax, points=polygon)
-                drawer.draw(
-                    corner_radius=corner_radius,
-                    facecolor=facecolor,
-                    edgecolor=color,
-                )
-
-                if kwds.get("legend", False):
-                    legend_info["handles"].append(
-                        plt.Rectangle(
-                            (0, 0),
-                            0,
-                            0,
-                            facecolor=facecolor,
-                            edgecolor=color,
-                        )
-                    )
-                    legend_info["labels"].append(str(color_id))
-
-            if kwds.get("legend", False):
-                ax.legend(
-                    legend_info["handles"],
-                    legend_info["labels"],
-                )
-
-        # Determine the order in which we will draw the vertices and edges
-        vertex_order = self._determine_vertex_order(graph, kwds)
-        edge_order = self._determine_edge_order(graph, kwds)
-
-        # Construct the iterator that we will use to draw the vertices
-        vs = graph.vs
-        if vertex_order is None:
-            # Default vertex order
-            vertex_coord_iter = zip(vs, vertex_builder, layout)
-        else:
-            # Specified vertex order
-            vertex_coord_iter = (
-                (vs[i], vertex_builder[i], layout[i]) for i in vertex_order
-            )
-
-        # Draw the vertices
-        drawer_method = vertex_drawer.draw
-        for vertex, visual_vertex, coords in vertex_coord_iter:
-            drawer_method(visual_vertex, vertex, coords)
-
-        # Construct the iterator that we will use to draw the vertex labels
-        vs = graph.vs
-        if vertex_order is None:
-            # Default vertex order
-            vertex_coord_iter = zip(vertex_builder, layout)
-        else:
-            # Specified vertex order
-            vertex_coord_iter = ((vertex_builder[i], layout[i]) for i in vertex_order)
-
-        # Draw the vertex labels
-        for vertex, coords in vertex_coord_iter:
-            if vertex.label is None:
-                continue
-
-            label_size = kwds.get(
-                "vertex_label_size",
-                vertex.label_size,
-            )
-
-            ax.text(
-                *coords,
-                vertex.label,
-                fontsize=label_size,
-                # TODO: alignment, overlap, offset, etc.
-            )
-
-        # Construct the iterator that we will use to draw the edges
-        es = graph.es
-        if edge_order is None:
-            # Default edge order
-            edge_coord_iter = zip(es, edge_builder)
-        else:
-            # Specified edge order
-            edge_coord_iter = ((es[i], edge_builder[i]) for i in edge_order)
-
-        # Draw the edges
-        if directed:
-            drawer_method = edge_drawer.draw_directed_edge
-        else:
-            drawer_method = edge_drawer.draw_undirected_edge
-        for edge, visual_edge in edge_coord_iter:
-            src, dest = edge.tuple
-            src_vertex, dest_vertex = vertex_builder[src], vertex_builder[dest]
-            drawer_method(visual_edge, src_vertex, dest_vertex)
-
-        # Draw the edge labels
-        labels = kwds.get("edge_label", None)
-        if labels is not None:
-            edge_label_iter = (
-                (labels[i], edge_builder[i], graph.es[i]) for i in range(graph.ecount())
-            )
-            for label, visual_edge, edge in edge_label_iter:
-                # Ask the edge drawer to propose an anchor point for the label
-                src, dest = edge.tuple
-                src_vertex, dest_vertex = vertex_builder[src], vertex_builder[dest]
-                (x, y), (halign, valign) = edge_drawer.get_label_position(
-                    edge,
-                    src_vertex,
-                    dest_vertex,
-                )
-
-                ax.text(
-                    x,
-                    y,
-                    label,
-                    fontsize=visual_edge.label_size,
-                    color=visual_edge.label_color,
-                    ha=halign,
-                    va=valign,
-                    # TODO: offset, etc.
-                )
-
-        # Despine
-        ax.spines["right"].set_visible(False)
-        ax.spines["top"].set_visible(False)
-        ax.spines["left"].set_visible(False)
-        ax.spines["bottom"].set_visible(False)
-
-        # Remove axis ticks
-        ax.set_xticks([])
-        ax.set_yticks([])
-
-        ax.autoscale_view()
